Working towards built-in updates for cell refs (WIP).
